% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blend.R
\name{exclusion}
\alias{exclusion}
\title{Blend colours and palettes using exclusion mode}
\usage{
exclusion(...)
}
\arguments{
\item{...}{Either one or two colour/palette arguments:
\itemize{
\item If one argument: the colour or palette is blended with itself
\item If two arguments: the first is blended with the second
Each argument can be a character vector of colours or a \verb{scales::pal_*()} function
}}
}
\description{
Blend colours and palettes using exclusion mode
}
\examples{
exclusion("#FFA600FF", "#8991A1FF")
}
