% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blend.R
\name{hard_light}
\alias{hard_light}
\title{Blend colours and palettes using hard light mode}
\usage{
hard_light(...)
}
\arguments{
\item{...}{Either one or two colour/palette arguments:
\itemize{
\item If one argument: the colour or palette is blended with itself
\item If two arguments: the first is blended with the second
Each argument can be a character vector of colours or a \verb{scales::pal_*()}
function
}}
}
\value{
Character vector of blended colours or a blending function.
}
\description{
Combines multiply and screen depending on the lightness of the first colour.
Like overlay but the first colour controls whether darkening or lightening
is applied.
}
\examples{
hard_light("#FFA600FF", "#8991A1FF")
}
