% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paletteblend-package.R
\docType{package}
\name{paletteblend-package}
\alias{paletteblend}
\alias{paletteblend-package}
\title{paletteblend: Blend Colours and Palettes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Blend colours and palettes using blend modes such as multiply and screen.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/davidhodge931/paletteblend}
  \item \url{https://davidhodge931.github.io/paletteblend/}
  \item Report bugs at \url{https://github.com/davidhodge931/paletteblend/issues}
}

}
\author{
\strong{Maintainer}: David Hodge \email{davidhodge931@gmail.com} (\href{https://orcid.org/0000-0002-3868-7501}{ORCID}) [copyright holder]

}
\keyword{internal}
