% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blend.R
\name{colour_burn}
\alias{colour_burn}
\title{Blend colours using colour burn mode}
\usage{
colour_burn(...)
}
\arguments{
\item{...}{Either one or two colour arguments:
\itemize{
\item If one argument: the colour is blended with itself
\item If two arguments: the first is blended with the second
Each argument can be a character vector of colours or a \verb{scales::pal_*()} function
}}
}
\description{
Blend colours using colour burn mode
}
\examples{
colour_burn("#FFA600FF", "#8991A1FF")
}
