% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blend.R
\name{hard_light}
\alias{hard_light}
\title{Blend colours using hard light mode}
\usage{
hard_light(...)
}
\arguments{
\item{...}{Either one or two colour arguments:
\itemize{
\item If one argument: the colour is blended with itself
\item If two arguments: the first is blended with the second
Each argument can be a character vector of colours or a \verb{scales::pal_*()} function
}}
}
\description{
Blend colours using hard light mode
}
\examples{
hard_light("#FFA600FF", "#8991A1FF")
}
